feat(node)!: request compressed responses by default
BREAKING CHANGE: if the server responded with a brotli-compressed response,
previously we would return the raw, compressed value. With this change, we
will return the decompressed one. This is a pretty _theoretical_ scenario.

BREAKING CHANGE: if the server somehow incorrectly handles `accept-encoding`,
this change might cause the request to fail. Again, this is a pretty unlikely
scenario.