test: don't depend on default keep-alive settings of Node
It appears that some OS (Ubuntu in CI at least) actually has keep-alive
set to true by default? This is a bit weird and is not documented
anywhere. This changes the agent tests to explicitly test that
keep-alive can be set to true/false and makes no assumption of
what happens without it.